some fix in _unmap() method

+++ networks/anchor_target_layer.py
@@ -60,11 +60,11 @@ class AnchorTargetLayer(object):
         """ Unmap a subset of item (data) back to the original set of items (of
         size count) """
         if len(data.shape) == 1:
-            ret = np.empty((count,), dtype=np.float32)
+            ret = np.empty((count,), dtype=data.dtype)
             ret.fill(fill)
             ret[inds] = data
         else:
-            ret = np.empty((count,) + data.shape[1:], dtype=np.float32)
+            ret = np.empty((count,) + data.shape[1:], dtype=data.dtype)
             ret.fill(fill)
             ret[inds, :] = data
         return ret
